requirements.txt
# Add your libraries here, e.g.:
# requests==2.31.0
requests
psycopg2-binary